Tenth commit
"Archive" directory added. This directory will contain directories
that correspond to videos from the playlist.

For example, a directory named "Ep3-FirstJqueryCode" would
correspond to this video from the playlist:
https://www.youtube.com/watch?v=y4Qc3LF8Pz0&list=PLr6-GrHUlVf_RNxQQkQnEwUiHELmB0fW1&index=4&t=44s

As of this commit, there is one directory named "V0-0". This directory
is simply the home of the base files before the tutorial has been
started.