Seventeenth commit
Changes to "reset.css":

box-sixing: border box; rule added to :root.

Changes to "stylesheet.css":

Rules for wrappers and footer added to fix issue with footer not
sticking to bottom of page.

Changes to "index.html":

Wrappers added to fix issue with footer not sticking to bottom of of
page.